Improve `everyInterval()` (#4441)
* Improve `everyInterval()`

The function `everyInterval()` was used to run all user's code in parallel. Since this doesn't work, when the execution time is bigger than the interval, it was changed to run the user's code and then pauses as long as the set interval. 

If the interval is 10 seconds and the execution of the code takes 5 seconds, the real interval between each code execution increases to 15s. If the user's code is 20 seconds, it's increasing to 30s. 

I changed this to measure the execution time of the user's code and only apply a pause if there is time left. For a chosen interval of 10 seconds and a script execution time of 5 this will end up to pause 5 additional seconds and the interval will be 10s as expected. If the execution time is 20 seconds it will not pause at all and the interval will be 20s since it just can't be shorter without causing issues.

* remove newline at end of file

* remove newline at end of file

* remove newline at end of file

* remove let now

Co-authored-by: JW <gitkraken@juriwolf.de>