Melody Block Updates to Align with Previous Blocks (#5228)
The goal is to align the new melody blocks added in #5123 (specifically the "built in melody" block) to something that resembles the previous blocks a bit more closely. See #5219 for more details on why.

Specific changes were to include "melody" before the melody name and to change the default to "in background". Since we don't want to change all blocks to use "in background" by default, I created a copy of the play block and used that instead. I don't love this approach, as it makes the javascript look pretty weird, but it gets the job done and a cleaner solution would require more time than we'd like to commit to this change.